Trc20 转账amount 从 long 改变成 bigInteger

+++ src/main/java/com/yan/wallet/chain/tron/bean/Trc20TransferParam.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigInteger;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class Trc20TransferParam {
     private String fromAddress;
     private String toAddress;
-    private Long amount;
+    private BigInteger amount;
     private String tokenAddress;
     private String memo;
 }
